requirements.txt
# PyTest
pyyaml

# Code check
pylint
flake8
flake8_docstrings
hacking
